Refactored Entity's list of components and associated methods into the ComponentList mixin class.
