+++ OSE V2/OSE V2.vcxproj
@@ -162,6 +162,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClInclude Include="OSE-Core\Entity\Component\ComponentList.h" />
     <ClInclude Include="OSE-Core\Rendering\RenderingFactory.h" />
     <ClInclude Include="OSE-Core\Math\ITransform.h" />
     <ClInclude Include="OSE-Core\Math\Transformable.h" />
@@ -647,6 +648,7 @@
     <ClInclude Include="OSE-V2-STD-Modules\EngineDependencies\glm\vector_relational.hpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="OSE-Core\Entity\Component\ComponentList.cpp" />
     <ClCompile Include="OSE-Core\EngineReferences.cpp" />
     <ClCompile Include="OSE-Core\Rendering\RenderingEngine.cpp" />
     <ClCompile Include="OSE-Core\Editor\Editor.cpp" />
