+++ OSE V2/OSE V2.vcxproj.filters
@@ -171,6 +171,9 @@
     <ClCompile Include="OSE-Core\Entity\Component\Component.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="OSE-Core\Entity\Component\ComponentList.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="OSE-V2-STD-Modules\EngineDependencies\glm\common.hpp">
@@ -1620,6 +1623,9 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="OSE-Core\Entity\Component\SpriteRenderer.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="OSE-Core\Entity\Component\ComponentList.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
